requirements.txt
audiomentations==0.34.1
numpy==1.23.5
pandas==1.5.3
pandas==2.0.1
scipy==1.11.4
torch==2.1.0
tqdm==4.65.0
python-dotenv
